Split utils functions in a new file
Add locking for writing on deployment object
