+++ jumpscale/packages/jukebox/bottle/jukebox.py
@@ -11,7 +11,7 @@ from jumpscale.core.base import StoredFactory
 from jumpscale.packages.auth.bottle.auth import authenticated, get_user_info, login_required, package_authorized
 
 from jumpscale.packages.jukebox.bottle.models import UserEntry
-from jumpscale.sals.jukebox import jukebox
+from jumpscale.sals.jukebox import utils
 
 app = Bottle()
 
@@ -116,7 +116,7 @@ def accept():
         user_entry.explorer_url = explorer_url
         user_entry.tname = tname
         user_entry.save()
-        jukebox.get_or_create_user_wallet(f"jukebox_{j.data.text.removesuffix(tname, '.3bot')}")
+        utils.get_or_create_user_wallet(f"jukebox_{j.data.text.removesuffix(tname, '.3bot')}")
         create_intermediate_identity(tname=tname, email=user_info["email"], explorer_url=explorer_url)
         return HTTPResponse(
             j.data.serializers.json.dumps({"allowed": True}), status=201, headers={"Content-Type": "application/json"}
@@ -133,7 +133,7 @@ def allowed():
     instances = user_factory.list_all()
     for name in instances:
         user_entry = user_factory.get(name)
-        jukebox.get_or_create_user_wallet(f"jukebox_{j.data.text.removesuffix(tname, '.3bot')}")
+        utils.get_or_create_user_wallet(f"jukebox_{j.data.text.removesuffix(tname, '.3bot')}")
         if user_entry.tname == tname and user_entry.explorer_url == explorer_url and user_entry.has_agreed:
             create_intermediate_identity(
                 tname=tname, email=user_info["email"], explorer_url=explorer_url
@@ -217,7 +217,7 @@ def get_wallet():
     tname = user_info["username"]
     prefixed_tname = f"{IDENTITY_PREFIX}_{tname.replace('.3bot', '')}"
 
-    data = jukebox.get_wallet_funding_info(prefixed_tname)
+    data = utils.get_wallet_funding_info(prefixed_tname)
     if not data:
         return HTTPResponse(
             j.data.serializers.json.dumps({"wallet": False}), status=404, headers={"Content-Type": "application/json"}
