+++ jumpscale/packages/jukebox/chats/ubuntu.py
@@ -24,13 +24,10 @@ class UbuntuDeploy(JukeboxDeployChatflow):
 
     @chatflow_step(title="User configurations")
     def environment(self):
+        self.env = {}
         # self.env = {"pub_key": self.public_key}
         self.metadata = {
-            "form_info": {
-                "chatflow": self.SOLUTION_TYPE,
-                "Solution name": self.deployment_name,
-                "number_of_nodes": self.nodes_count,
-            },
+            "form_info": {"chatflow": self.SOLUTION_TYPE, "Solution name": self.deployment_name,},
         }
         self.secret_env = {}
