+++ jumpscale/sals/jukebox/jukebox.py
@@ -3,26 +3,21 @@ from time import sleep
 import uuid
 
 import gevent
-from jumpscale.clients.explorer.models import Container, DiskType, NextAction, WorkloadType
-from jumpscale.clients.explorer.models import DiskType, State as WorkloadState
-from jumpscale.clients.stellar import TRANSACTION_FEES
+from jumpscale.clients.explorer.models import DiskType, State as WorkloadState, NextAction
 from jumpscale.core.base import Base, fields
 from jumpscale.core.base import Base, fields
 from jumpscale.loader import j
 from jumpscale.sals.reservation_chatflow import DeploymentFailed, deployer
 
+from jumpscale.sals.jukebox import utils
 from jumpscale.sals.jukebox.models import BlockchainNode, State
-from jumpscale.sals.vdc.scheduler import GlobalCapacityChecker, Scheduler
+from jumpscale.sals.vdc.scheduler import Scheduler
+from gevent.lock import BoundedSemaphore
 
-# 1. create pool
-# 2. create network
-# 3. create wallet for user
-# 4. create indentity for the user
-# 5. blockchain ndoes crud
 
 CURRENCIES = ["TFT"]
 IDENTITY_PREFIX = "jukebox"
-JUKEBOX_AUTO_EXTEND_KEY = "jukebox:auto_extend"
+POOL_EXPIRATION_VALUE = 9223372036854775807
 
 
 def on_exception(greenlet_thread):
@@ -33,150 +28,6 @@ def on_exception(greenlet_thread):
     message = f"raised an exception: {greenlet_thread.exception}"
     j.tools.alerthandler.alert_raise(app_name="jukebox", message=message, alert_type="exception")
     j.logger.error(message)
-
-
-# flist_map = {
-#     "digibyte": {"flist": "http"},
-#     "dash": "",
-#     "matic": "",
-#     "ubuntu": "https://hub.grid.tf/tf-bootable/3bot-ubuntu-20.04.flist",
-#     "presearch": "https://hub.grid.tf/waleedhammam.3bot/arrajput-presearch-latest.flist",
-# }
-
-
-def get_or_create_user_wallet(wallet_name):
-    # Create a wallet for the user to be used in extending his pool
-    if not j.clients.stellar.find(wallet_name):
-        wallet = j.clients.stellar.new(wallet_name)
-        try:
-            wallet.activate_through_activation_wallet()
-        except Exception:
-            j.clients.stellar.delete(name=wallet_name)
-            raise j.exceptions.JSException("Error on wallet activation")
-
-        try:
-            wallet.add_known_trustline("TFT")
-        except Exception:
-            j.clients.stellar.delete(name=wallet_name)
-            raise j.exceptions.JSException(
-                f"Failed to add trustlines to wallet {wallet_name}. Any changes made will be reverted."
-            )
-
-        wallet.save()
-    else:
-        wallet = j.clients.stellar.get(wallet_name)
-    return wallet
-
-
-def calculate_payment_from_container_resources(
-    cpu, memory, disk_size, duration, farm_id=None, farm_name="freefarm", disk_type=DiskType.SSD
-):
-    if farm_name and not farm_id:
-        zos = j.sals.zos.get()
-        farm_id = zos._explorer.farms.get(farm_name=farm_name).id
-    empty_container = Container()
-    empty_container.capacity.cpu = cpu
-    empty_container.capacity.memory = memory
-    empty_container.capacity.disk_size = disk_size
-    empty_container.capacity.disk_type = disk_type
-    cost = j.tools.zos.consumption.cost(empty_container, duration=duration, farm_id=farm_id)
-    return cost
-
-
-def calculate_required_units(cpu, memory, disk_size, duration_seconds, number_of_containers=1):
-    cont = Container()
-    cont.capacity.cpu = cpu
-    cont.capacity.memory = memory
-    cont.capacity.disk_size = disk_size
-    cont.capacity.disk_type = DiskType.SSD
-
-    cloud_units = {"cu": 0, "su": 0, "ipv4u": 0}
-    cont_units = cont.resource_units().cloud_units()
-    cloud_units["cu"] += cont_units.cu * number_of_containers
-    cloud_units["su"] += cont_units.su * number_of_containers
-    cloud_units["cu"] *= duration_seconds
-    cloud_units["su"] *= duration_seconds
-    return cloud_units
-
-
-def get_possible_farms(cru, sru, mru, number_of_deployments):
-    gcc = GlobalCapacityChecker()
-    farm_names = gcc.get_available_farms(
-        cru=cru * number_of_deployments,
-        mru=mru * number_of_deployments,
-        sru=sru * number_of_deployments,
-        # ip_version=None,
-        # no_nodes=1,
-        accessnodes=True,
-    )
-    return farm_names
-
-
-def get_network_ip_range():
-    return j.sals.reservation_chatflow.reservation_chatflow.get_ip_range()
-
-
-def show_payment(bot, cost, wallet_name, expiry=5, description=None):
-    payment_id, _ = j.sals.billing.submit_payment(
-        amount=cost, wallet_name=wallet_name, refund_extra=False, expiry=expiry, description=description
-    )
-
-    if cost > 0:
-        notes = []
-        return j.sals.billing.wait_payment(payment_id, bot=bot, notes=notes), cost, payment_id
-    return True, cost, payment_id
-
-
-def calculate_funding_amount(identity_name):
-    identity = j.core.identity.find(identity_name)
-    zos = j.sals.zos.get(identity_name)
-    if not identity:
-        return 0
-    total_price = 0
-    deployments = j.sals.jukebox.list(identity_name=identity_name)
-    for deployment in deployments:
-        price = 0
-        if not deployment.auto_extend:
-            continue
-        if deployment.expiration_date.timestamp() > j.data.time.utcnow().timestamp + 60 * 60 * 24 * 2:
-            continue
-        pool_id = deployment.pool_ids[0]  # TODO change when having multiple pools
-        pool = zos.pools.get(pool_id)  # TODO: set active unit while listing
-        cus = pool.active_cu
-        sus = pool.active_su
-        ipv4us = pool.active_ipv4
-        farm_id = zos._explorer.farms.get(farm_name=deployment.farm_name).id
-        farm_prices = zos._explorer.farms.get_deal_for_threebot(farm_id, j.core.identity.me.tid)[
-            "custom_cloudunits_price"
-        ]
-        price += zos._explorer.prices.calculate(cus=cus, sus=sus, ipv4us=ipv4us, farm_prices=farm_prices)
-        price *= 60 * 60 * 24 * 30
-        price += TRANSACTION_FEES
-        total_price += price
-    return total_price
-
-
-def get_wallet_funding_info(identity_name):
-    wallet = j.clients.stellar.find(identity_name)
-    if not wallet:
-        return {}
-
-    asset = "TFT"
-    current_balance = wallet.get_balance_by_asset(asset)
-    amount = calculate_funding_amount(identity_name) - current_balance
-    amount = 0 if amount < 0 else round(amount, 6)
-
-    qrcode_data = f"TFT:{wallet.address}?amount={amount}&message=topup&sender=me"
-    qrcode_image = j.tools.qrcode.base64_get(qrcode_data, scale=3)
-
-    data = {
-        "address": wallet.address,
-        "balance": {"amount": current_balance, "asset": asset},
-        "amount": amount,
-        "qrcode": qrcode_image,
-        "network": wallet.network.value,
-    }
-    return data
 
 
 class JukeboxDeployment(Base):
@@ -195,12 +46,27 @@ class JukeboxDeployment(Base):
     disk_size = fields.Integer()  # per node
     disk_type = fields.Enum(DiskType)  # per node
     _zos = None
+    __lock = BoundedSemaphore(1)
 
     @property
     def zos(self):
         if self._zos is None:
             self._zos = j.sals.zos.get(self.identity_name)
         return self._zos
+
+    def lock_deployment(function):
+        def wrapper(self, *args, **kwargs):
+            self.__lock.acquire()
+            try:
+                result = function(self, *args, **kwargs)
+            except Exception as e:
+                raise e
+            finally:
+                self.__lock.release()
+
+            return result
+
+        return wrapper
 
     def create_capacity_pool(self, wallet, cu=100, su=100, ipv4us=0, farm="freefarm"):
         payment_detail = self.zos.pools.create(cu=cu, su=su, ipv4us=ipv4us, farm=farm)
@@ -214,7 +80,7 @@ class JukeboxDeployment(Base):
         self.pool_ids.append(payment_detail.reservation_id)
         return payment_detail.reservation_id
 
-    def get_container_ip(self, network_name, node, pool_id, tname, excluded_ips=None):
+    def get_container_ip(self, network_name, node, excluded_ips=None):
         excluded_ips = excluded_ips or []
         excluded_nodes = j.core.db.get("excluded_nodes")
         if not excluded_nodes:
@@ -225,7 +91,12 @@ class JukeboxDeployment(Base):
         network_view = deployer.get_network_view(network_name, identity_name=self.identity_name)
         network_view_copy = network_view.copy()
         result = deployer.add_network_node(
-            network_view.name, node, pool_id, network_view_copy, identity_name=self.identity_name, owner=tname
+            network_view.name,
+            node,
+            self.pool_ids[0],
+            network_view_copy,
+            identity_name=self.identity_name,
+            owner=self.identity_name,
         )
 
         if result:
@@ -248,16 +119,16 @@ class JukeboxDeployment(Base):
             if ip not in excluded_ips:
                 return ip
 
-    def deploy_network(self, pool_id, network_name, owner_tname, ip_range=None):
-        ip_range = ip_range or get_network_ip_range()
-        scheduler = Scheduler(pool_id=pool_id)
+    def deploy_network(self, network_name, ip_range=None):
+        ip_range = ip_range or utils.get_network_ip_range()
+        scheduler = Scheduler(pool_id=self.pool_ids[0])
         network_success = False
         ip_version = "IPv4"
         for access_node in scheduler.nodes_by_capacity(ip_version=ip_version, accessnodes=True):
             j.logger.info(f"Deploying network on node {access_node.node_id}")
             network_success = True
             result = deployer.deploy_network(
-                network_name, access_node, ip_range, ip_version, pool_id, self.identity_name
+                network_name, access_node, ip_range, ip_version, self.pool_ids[0], self.identity_name
             )
             for wid in result["ids"]:
                 try:
@@ -277,31 +148,17 @@ class JukeboxDeployment(Base):
             if network_success:
                 # store wireguard config
                 j.logger.info(
-                    f"saving wireguard config to {j.core.dirs.CFGDIR}/jukebox/wireguard/{owner_tname}/{network_name}.conf"
+                    f"saving wireguard config to {j.core.dirs.CFGDIR}/jukebox/wireguard/{self.identity_name}/{network_name}.conf"
                 )
                 wg_quick = result["wg"]
-                j.sals.fs.mkdirs(f"{j.core.dirs.CFGDIR}/jukebox/wireguard/{owner_tname}")
+                j.sals.fs.mkdirs(f"{j.core.dirs.CFGDIR}/jukebox/wireguard/{self.identity_name}")
                 j.sals.fs.write_file(
-                    f"{j.core.dirs.CFGDIR}/jukebox/wireguard/{owner_tname}/{network_name}.conf", wg_quick
+                    f"{j.core.dirs.CFGDIR}/jukebox/wireguard/{self.identity_name}/{network_name}.conf", wg_quick
                 )
                 return True, wg_quick
 
     def deploy_all_containers(
-        self,
-        farm_name,
-        number_of_deployments,
-        network_name,
-        cru,
-        sru,
-        mru,
-        pool_ids,
-        owner_tname,
-        blockchain_type,
-        env=None,
-        secret_env=None,
-        metadata=None,
-        flist=None,
-        entry_point="",
+        self, number_of_deployments, network_name, env=None, secret_env=None, metadata=None, flist=None, entry_point="",
     ):
         metadata = metadata or {}
         env = env or {}
@@ -313,78 +170,51 @@ class JukeboxDeployment(Base):
             excluded_nodes = []
         else:
             excluded_nodes = j.data.serializers.json.loads(excluded_nodes.decode())
-        scheduler = Scheduler(farm_name=farm_name)
+        scheduler = Scheduler(farm_name=self.farm_name)
         scheduler.exclude_nodes(*excluded_nodes)
         deployment_threads = []
-        for i in range(number_of_deployments):
+        i = 0
+        while i < number_of_deployments:
             # for each node check how many containers can be deployed on it, and based on that assign that node for X deployments
-            pool_id = pool_ids[0]  # TODO
-            node = next(scheduler.nodes_by_capacity(cru=cru, sru=sru, mru=mru))
+            node = next(scheduler.nodes_by_capacity(cru=self.cpu, sru=self.disk_size / 1024, mru=self.memory / 1024))
 
             excluded_ips = used_ip_addresses.get(node.node_id, [])
-            ip_address = self.get_container_ip(network_name, node, pool_id, owner_tname, excluded_ips)
+            ip_address = self.get_container_ip(network_name, node, excluded_ips)
             if not ip_address:
-                i -= 1
                 continue
 
             used_ip_addresses[node.node_id].append(ip_address)
 
             # START SPAWN
-
             thread = gevent.spawn(
-                self.deploy_container,
-                network_name,
-                node,
-                pool_id,
-                ip_address,
-                blockchain_type,
-                env,
-                metadata,
-                flist,
-                entry_point,
-                secret_env,
-                cru,
-                sru * 1024,
-                mru * 1024,
+                self.deploy_container, network_name, node, ip_address, env, metadata, flist, entry_point, secret_env,
             )
             thread.link_exception(on_exception)
             deployment_threads.append(thread)
+            i += 1
             # END SPAWN
         # TODO check resv ids success/failure, if resv failed retry on same node
         gevent.joinall(deployment_threads)
 
     def deploy_container(
-        self,
-        network_name,
-        node,
-        pool_id,
-        ip_address,
-        blockchain_type="ubuntu",
-        env=None,
-        metadata=None,
-        flist=None,
-        entry_point="",
-        secret_env=None,
-        cpu=1,
-        memory=1024,
-        disk_size=512,
+        self, network_name, node, ip_address, env=None, metadata=None, flist=None, entry_point="", secret_env=None,
     ):
         metadata = metadata or {}
         env = env or {}
         secret_env = secret_env or {}
         if not flist:
-            raise Exception(f"Flist for {blockchain_type} not found, Please pass it")
+            raise Exception(f"Flist for {self.solution_type} not found, Please pass it")
 
         resv_id = deployer.deploy_container(
             identity_name=self.identity_name,
-            pool_id=pool_id,
+            pool_id=self.pool_ids[0],
             node_id=node.node_id,
             network_name=network_name,
             ip_address=ip_address,
             flist=flist,
-            cpu=cpu,
-            memory=memory,
-            disk_size=disk_size,
+            cpu=self.cpu,
+            memory=self.memory,
+            disk_size=self.disk_size,
             env=env,
             interactive=False,
             entrypoint=entry_point,
@@ -416,6 +246,7 @@ class JukeboxDeployment(Base):
         self.save()
         return resv_id
 
+    @lock_deployment
     def delete_node(self, wid):
         for node in self.nodes:
             if node.wid == wid:
@@ -423,3 +254,44 @@ class JukeboxDeployment(Base):
                 self.nodes_count -= 1
                 self.save()
                 self.zos.workloads.decomission(node.wid)
+
+    # def redeploy_containers(self, number_of_containers):
+    #     wid = self.nodes[0].wid
+    #     workload = self.zos.workloads.get(wid)
+    #     network_name = f"{self.identity_name}_{self.pool_id}"
+    #     cru = self.cpu
+    #     mru = self.memory * 1024
+    #     sru = self.disk_size * 1024
+    #     env = workload.environment
+
+    #     self.deploy_all_containers(
+    #         network_name=network_name,
+    #         cru=cru,
+    #         mru=mru,
+    #         sru=sru,
+    #         pool_ids=self.pool_ids,
+    #         owner_tname=self.identity_name,
+    #         blockchain_type=self.deployment_name
+    #         env=workload.environment,
+    #         secret_env=
+    #         )
+
+    @lock_deployment
+    def _update_deployment(self):
+        pool = self.zos.pools.get(self.pool_ids[0])
+        if pool.empty_at == POOL_EXPIRATION_VALUE and pool.cus == 0:  # Also add check on sus if VMs are used
+            self.state = State.EXPIRED
+        elif pool.empty_at == POOL_EXPIRATION_VALUE:
+            self.state = State.ERROR
+        else:
+            self.expiration_date = pool.empty_at
+        for node in self.nodes:
+            workload = self.zos.workloads.get(node.wid)
+            if workload.info.next_action == NextAction.DEPLOY or node.state in [State.DELETED, State.EXPIRED]:
+                continue
+
+            elif pool.empty_at == POOL_EXPIRATION_VALUE and pool.cus == 0:  # Also add check on sus if VMs are used
+                node.state = State.EXPIRED
+            else:
+                node.state = State.ERROR
+        self.save()
